+++ src/config.rs
@@ -808,6 +808,22 @@ pub enum NotchWidth {
     Bandwidth { freq: PrcFmt, bandwidth: PrcFmt },
 }
 
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+#[serde(deny_unknown_fields)]
+pub struct GeneralNotchParams {
+    pub freq_p: PrcFmt,
+    pub freq_z: PrcFmt,
+    pub q_p: PrcFmt,
+    #[serde(default)]
+    pub normalize_at_dc: Option<bool>,
+}
+
+impl GeneralNotchParams {
+    pub fn normalize_at_dc(&self) -> bool {
+        self.normalize_at_dc.unwrap_or_default()
+    }
+}
+
 #[allow(clippy::upper_case_acronyms)]
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 #[serde(tag = "type")]
@@ -851,6 +867,7 @@ pub enum BiquadParameters {
     },
     Bandpass(NotchWidth),
     Notch(NotchWidth),
+    GeneralNotch(GeneralNotchParams),
     LinkwitzTransform {
         freq_act: PrcFmt,
         q_act: PrcFmt,
@@ -879,6 +896,9 @@ pub enum BiquadComboParameters {
         freq: PrcFmt,
         order: usize,
     },
+    Tilt {
+        gain: PrcFmt,
+    },
     FivePointPeq {
         fls: PrcFmt,
         qls: PrcFmt,
@@ -896,6 +916,27 @@ pub enum BiquadComboParameters {
         qhs: PrcFmt,
         ghs: PrcFmt,
     },
+    GraphicEqnalizer(GraphicEqualizerParameters),
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+#[serde(deny_unknown_fields)]
+pub struct GraphicEqualizerParameters {
+    #[serde(default)]
+    freq_min: Option<f32>,
+    #[serde(default)]
+    freq_max: Option<f32>,
+    pub gains: Vec<f32>,
+}
+
+impl GraphicEqualizerParameters {
+    pub fn freq_min(&self) -> f32 {
+        self.freq_min.unwrap_or(20.0)
+    }
+
+    pub fn freq_max(&self) -> f32 {
+        self.freq_max.unwrap_or(20000.0)
+    }
 }
 
 #[derive(Clone, Copy, Debug, Serialize, Deserialize, Eq, PartialEq)]